feat(laravel api backend): :sparkles: add storage/app in .gitignore at laravel api

+++ app/Http/Controllers/UserController.php
@@ -2,11 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\{User, Role};
+use Illuminate\Http\Request;
+use App\Http\Resources\UserResource;
 use App\Exceptions\NotFoundException;
 use App\Http\Requests\User\StoreUserRequest;
 use App\Http\Requests\User\UpdateUserRequest;
-use App\Http\Resources\UserResource;
-use App\Models\User;
 use Gate;
 use Str;
 
@@ -14,7 +15,7 @@ class UserController extends Controller
 {
     public function index()
     {
-        // Gate::authorize('list_users', User::class);
+        Gate::authorize('list_users', User::class);
         $users = User::paginate();
         return UserResource::collection($users);
     }
@@ -22,7 +23,7 @@ class UserController extends Controller
     public function show(int $userId)
     {
         $user = $this->user($userId);
-        // Gate::authorize('show_user', $user);
+        Gate::authorize('show_user', $user);
         return UserResource::make($user);
     }
     
@@ -40,14 +41,14 @@ class UserController extends Controller
         }
 
         $user   = User::create($input);
-        // Gate::authorize('update_user', $user);
+        Gate::authorize('update_user', $user);
         return UserResource::make($user);   
     }
 
     public function update(int $userId, UpdateUserRequest $request)
     {
         $user = $this->user($userId);
-        // Gate::authorize('update_user', $user);
+        Gate::authorize('update_user', $user);
         $user->update($request->validated());
         return UserResource::make($user);   
     }
@@ -55,9 +56,16 @@ class UserController extends Controller
     public function destroy(int $userId)
     {
         $user = $this->user($userId);
-        // Gate::authorize('destroy_user', $user);
+        Gate::authorize('destroy_user', $user);
         $user->delete();
         return response()->noContent(); 
+    }
+
+    public function setRole(User $user, Request $request):void
+    {
+        Gate::authorize('update_user', $user);
+        $role = Role::where('name', $request->name)->first();
+        $user->roles()->attach($role);
     }
     
     private function user($userId)
